feat: make graphql collector modular
