+++ .github/actions/go/action.yaml
@@ -3,7 +3,7 @@ description: "This action install go and cache modules and build dependencies"
 inputs:
   go-version:
     description: "The go version to install"
-    default: "1.21"
+    default: "1.22"
     required: false
   cache-dependency-path:
     description: "The path to the dependency to cache"
