+++ go.work
@@ -1,4 +1,4 @@
-go 1.21.5
+go 1.22.0
 
 toolchain go1.22.0
