+++ graphqlmetrics/Dockerfile
@@ -1,4 +1,4 @@
-FROM --platform=${BUILDPLATFORM} golang:1.21 AS builder
+FROM --platform=${BUILDPLATFORM} golang:1.22 AS builder
 
 ARG TARGETOS
 ARG TARGETARCH
