+++ graphqlmetrics/core/server.go
@@ -2,198 +2,88 @@ package core
 
 import (
 	"context"
-	"errors"
-	"net/http"
-	"time"
-
-	"connectrpc.com/connect"
-	"github.com/prometheus/client_golang/prometheus"
+	"fmt"
+	"github.com/KimMachineGun/automemlimit/memlimit"
+	"github.com/dustin/go-humanize"
 	"github.com/wundergraph/cosmo/graphqlmetrics/gen/proto/wg/cosmo/graphqlmetrics/v1/graphqlmetricsv1connect"
+	"github.com/wundergraph/cosmo/graphqlmetrics/internal/logging"
+	"github.com/wundergraph/cosmo/graphqlmetrics/internal/migration"
+	"github.com/wundergraph/cosmo/graphqlmetrics/pkg/config"
 	"github.com/wundergraph/cosmo/graphqlmetrics/pkg/telemetry"
-	"go.opentelemetry.io/otel/attribute"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+	"go.uber.org/automaxprocs/maxprocs"
 	"go.uber.org/zap"
-	brotli "go.withmatt.com/connect-brotli"
+	"os"
 )
 
-// Option defines the method to customize Server.
-type Option func(s *Server)
-
-type Server struct {
-	server         *http.Server
-	listenAddr     string
-	logger         *zap.Logger
-	jwtSecret      []byte
-	metricsService graphqlmetricsv1connect.GraphQLMetricsServiceHandler
-
-	metricConfig     *telemetry.Config
-	prometheusServer *http.Server
-	registry         *prometheus.Registry
-
-	meterProvider *sdkmetric.MeterProvider
-}
-
-func NewServer(ctx context.Context, metricsService graphqlmetricsv1connect.GraphQLMetricsServiceHandler, opts ...Option) *Server {
-	s := &Server{
-		metricsService: metricsService,
-		listenAddr:     ":4005",
-		logger:         zap.NewNop(),
+func NewServer(ctx context.Context, cfg *config.Config, ms graphqlmetricsv1connect.GraphQLMetricsServiceHandler) (*Collector, error) {
+	cfg, err := config.LoadConfig()
+	if err != nil {
+		return nil, fmt.Errorf("could not load config: %w", err)
 	}
 
-	for _, opt := range opts {
-		opt(s)
+	logLevel, err := logging.ZapLogLevelFromString(cfg.LogLevel)
+	if err != nil {
+		return nil, fmt.Errorf("could not parse log level: %w", err)
 	}
 
-	s.bootstrap(ctx)
-
-	return s
-}
-
-func (s *Server) bootstrap(ctx context.Context) {
-	mux := http.NewServeMux()
-	var interceptors = []connect.Interceptor{}
-
-	if s.metricConfig.IsEnabled() {
-		mp, registry, err := s.metricConfig.NewPrometheusMeterProvider(ctx)
-		if err != nil {
-			s.logger.Error("Failed to create Prometheus exporter", zap.Error(err))
-		}
-		s.meterProvider = mp
-		s.registry = registry
-
-		baseAttributes := []attribute.KeyValue{}
-
-		metricsStore, err := telemetry.NewPromMetricStore(s.logger, s.meterProvider, baseAttributes)
-		if err != nil {
-			s.logger.Error("Failed to create metric store", zap.Error(err))
-		}
-
-		s.metricConfig.MetricStore = metricsStore
-	}
-
-	if s.metricConfig.Prometheus.Enabled {
-		s.prometheusServer = telemetry.NewPrometheusServer(s.logger, s.metricConfig.Prometheus.ListenAddr, s.metricConfig.Prometheus.Path, s.registry)
-		interceptors = append(
-			interceptors,
-			s.metricConfig.ObservabilityInterceptor(),
+	isDebug := cfg.LogLevel == "debug"
+	logger := logging.New(!cfg.JSONLog, isDebug, logLevel).
+		With(
+			zap.String("component", "@wundergraph/graphqlmetrics"),
+			zap.String("service_version", Version),
 		)
+
+	// Automatically set GOMAXPROCS to avoid CPU throttling on containerized environments
+	_, err = maxprocs.Set(maxprocs.Logger(logger.Sugar().Debugf))
+	if err != nil {
+		return nil, fmt.Errorf("could not set max GOMAXPROCS: %w", err)
 	}
 
-	path, handler := graphqlmetricsv1connect.NewGraphQLMetricsServiceHandler(
-		s.metricsService,
-		// Brotli compression support.
-		brotli.WithCompression(),
-		connect.WithInterceptors(interceptors...),
+	// Automatically set GOMEMLIMIT to 90% of the available memory.
+	// This is an effort to prevent the router from being killed by OOM (Out Of Memory)
+	// when the system is under memory pressure e.g. when GC is not able to free memory fast enough.
+	// More details: https://tip.golang.org/doc/gc-guide#Memory_limit
+	mLimit, err := memlimit.SetGoMemLimitWithOpts(
+		memlimit.WithRatio(0.9),
+		memlimit.WithProvider(
+			memlimit.ApplyFallback(
+				memlimit.FromCgroupHybrid,
+				memlimit.FromSystem,
+			),
+		),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("could not set memory limit: %w", err)
+	}
+	if mLimit > 0 {
+		logger.Info("GOMEMLIMIT set automatically", zap.String("limit", humanize.Bytes(uint64(mLimit))))
+	} else if os.Getenv("GOMEMLIMIT") != "" {
+		logger.Info("GOMEMLIMIT set by user", zap.String("limit", os.Getenv("GOMEMLIMIT")))
+	}
+
+	// Database migrations
+
+	if err := migration.Migrate(cfg.ClickHouseDSN, logger); err != nil {
+		return nil, fmt.Errorf("could not migrate database: %w", err)
+	}
+
+	metricsConfig := telemetry.NewTelemetryConfig(
+		Version,
+		telemetry.PrometheusConfig{
+			Enabled:    cfg.IsPrometheusEnabled,
+			ListenAddr: cfg.PrometheusListenAddr,
+			Path:       cfg.PrometheusPath,
+		},
 	)
 
-	healthHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
+	svr := NewCollector(
+		ctx,
+		ms,
+		WithJwtSecret([]byte(cfg.IngestJWTSecret)),
+		WithListenAddr(cfg.ListenAddr),
+		WithLogger(logger),
+		WithMetrics(metricsConfig),
+	)
 
-	mux.Handle("/health", healthHandler)
-	mux.Handle(path, authenticate(s.jwtSecret, s.logger, handler))
-
-	s.server = &http.Server{
-		Addr: s.listenAddr,
-		// https://ieftimov.com/posts/make-resilient-golang-net-http-servers-using-timeouts-deadlines-context-cancellation/
-		ReadTimeout:       1 * time.Minute,
-		WriteTimeout:      2 * time.Minute,
-		ReadHeaderTimeout: 20 * time.Second,
-		Handler:           mux,
-		ErrorLog:          zap.NewStdLog(s.logger),
-	}
-}
-
-func (s *Server) Start() error {
-	if s.server == nil {
-		return errors.New("server not initialized")
-	}
-
-	go func() {
-		if err := s.server.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			s.logger.Error("Could not start server", zap.Error(err))
-		}
-	}()
-
-	if s.metricConfig.Prometheus.Enabled {
-		go func() {
-			if err := s.startPrometheusServer(); err != nil && !errors.Is(err, http.ErrServerClosed) {
-				s.logger.Error("Could not start prometheus server", zap.Error(err))
-			}
-		}()
-	}
-
-	return nil
-}
-
-func (s *Server) Shutdown(ctx context.Context) error {
-	if s.server == nil {
-		return errors.New("server not initialized")
-	}
-
-	if err := s.server.Shutdown(ctx); err != nil {
-		s.logger.Error("Could not shutdown server", zap.Error(err))
-	}
-
-	if s.meterProvider != nil {
-		if err := s.meterProvider.Shutdown(ctx); err != nil {
-			s.logger.Error("Error shutting down meter provider", zap.Error(err))
-		}
-	}
-
-	if s.prometheusServer != nil {
-		if err := s.shutdownPrometheusServer(ctx); err != nil {
-			s.logger.Error("Could not shutdown prometheus server", zap.Error(err))
-		}
-	}
-
-	return nil
-}
-
-func (s *Server) shutdownPrometheusServer(ctx context.Context) error {
-	if s.prometheusServer == nil {
-		return errors.New("prometheus server was not initialized")
-	}
-
-	if err := s.prometheusServer.Shutdown(ctx); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (s *Server) startPrometheusServer() error {
-	if s.prometheusServer == nil {
-		return errors.New("prometheus server was not initialized")
-	}
-
-	if err := s.prometheusServer.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
-		return err
-	}
-
-	return nil
-}
-
-func WithLogger(logger *zap.Logger) Option {
-	return func(s *Server) {
-		s.logger = logger
-	}
-}
-
-func WithListenAddr(addr string) Option {
-	return func(s *Server) {
-		s.listenAddr = addr
-	}
-}
-
-func WithJwtSecret(secret []byte) Option {
-	return func(s *Server) {
-		s.jwtSecret = secret
-	}
-}
-
-func WithMetrics(cfg *telemetry.Config) Option {
-	return func(s *Server) {
-		s.metricConfig = cfg
-	}
+	return svr, nil
 }
